% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makedat.R
\name{makedat}
\alias{makedat}
\title{Set up the catch and covariate data as anomalies.  It takes a covariate name.}
\usage{
makedat(region = "Kerala", covname = "Precip", boxes = 1:14,
  rm.trend = TRUE, rm.season = TRUE, startyear = 1956, endyear = 2016,
  lag = 0, yearlag = 0, cov.qtr = 1:4)
}
\arguments{
\item{region}{"Kerala", "Karnataka", "Goa" or "SWCoast"}

\item{covname}{name of the covariate.  Must be a colname in seio_covariates_qtr.  If one of the covariates is from the boxes, then specify the part of the name in front of the number.}

\item{boxes}{for covariate from a box, the box number or numbers to use}

\item{rm.trend}{whether to detrend the catch and covariates}

\item{rm.season}{whether to deseason the catch and covariates}

\item{startyear}{first year}

\item{endyear}{last year}

\item{lag}{lag to use for the covariate. Do not set both this and yearlag.}

\item{yearlag}{If 0, it means covariate in year when sardine year starts; a sardine year starts in Q3.}

\item{cov.qtr}{the quarter to use for the covariate IF a specific quarter is wanted for the covariate.}
}
\value{
data frame
}
\description{
Creates a data frame for the demeaned LOG Kerala catch anomalies and standardized covariate anomolies.
You can specify that covariate value is lagged by 'lag' quarters or use
a specific quarter from a specific year as the covariate for all qtrs in a 'sardine year'.
A sardine year starts in Q3 and goes to Q2 the next calendar year.
}
\examples{
#covariate is rain in qtr 2 of the previous year
#my numbering is weird.  yearlag=1 & cov.qtr=2 would have the starred Q2 as the cov for the year in > <
Q1 *Q2* Q3 Q4 Q1 Q2 | >Q3 Q4 Q1 Q2< |
makedat(covname="Precip.Kerala", boxes=4, rm.trend=TRUE, rm.season=FALSE, yearlag=1, cov.qtr=2)

#yearlag=0 & cov.qtr=2 would have the starred Q2 as the cov for the sardine year in > <
Q1 Q2 Q3 Q4 Q1 *Q2* | >Q3 Q4 Q1 Q2< |
makedat(covname="Precip.Kerala", boxes=4, rm.trend=TRUE, rm.season=FALSE, yearlag=0, cov.qtr=2)
}
