% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{precip}
\alias{precip}
\title{Precipitation data}
\format{A data frame with:
\describe{
  \item{Year}{}
  \item{Month}{}
  \item{precip.gpcp.kerala}{Monthly GPCP precipitation in the box off the Kerala coast.}
  \item{precip.gpcp.karnataka}{Monthly GPCP precipitation in the box off the Karnataka coast.}
  \item{PrecipTRMM#}{Monthly TRMM precipitation in the boxes used for the other satellite data.}
  \item{Precip.Kerala}{Monthly precipitation from the land gauges.}
}}
\usage{
precip
}
\description{
Three precipition datasets off the SW coast of India.  Two are satellite derived and one is based on land gauges.
}
\details{
GPCP Precipitation data provided by the NOAA/NCEI Global Precipitation Climatology Project from their Web site at \url{https://www.ncei.noaa.gov/data/global-precipitation-climatology-project-gpcp-monthly}.  The data are on a 2.5 x 2.5 grid.  Two boxes are defined, one off the Kerala coast and one off the Karnataka coast.
\preformatted{
The boxes are 
Kerala Lat(8.75,  11.25), Lon(73.25,  75.75)
Karnataka Lat(13.75,  16.25), Lon(71.25,  73.75)
}
}
\keyword{datasets}
